- Created separate reducer, selector, effect and folders for subscription logic
