+++ src/app/auth/types/authState.interface.ts
@@ -6,5 +6,4 @@ export interface AuthStateInterface {
   currentUser: CurrentUserInterface | null
   isLoggedIn: boolean | null
   validationErrors: BackendErrorsInterface | null
-  isSubscribed: boolean
 }
