+++ src/app/general/types/subscriptionState.interface.ts
@@ -1,0 +1,3 @@
+export interface SubscriptionStateInterface {
+  isSubscribed: boolean
+}
