+++ src/app/shared/types/appState.interface.ts
@@ -1,5 +1,7 @@
 import {AuthStateInterface} from '../../auth/types/authState.interface'
+import {SubscriptionStateInterface} from '@app/general/types/subscriptionState.interface'
 
 export interface AppStateInterface {
-  auth: AuthStateInterface
+  auth: AuthStateInterface,
+  subscription: SubscriptionStateInterface
 }
